# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AGENDA TELEFÔNICA",
        "description": "API para gerenciamento de AGENDA TELEFÔNICA",
        "version": "1.0.0"
    },
    "paths": {
        "/api": {},
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Retorna todos os contatos registrados",
                "operationId": "bc39494d59500e2d75e6f6648b43c8de",
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Cria um novo contato",
                "operationId": "779a1f2268bae1138e7ea7843dd88d90",
                "requestBody": {
                    "description": "Dados do contato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "birth_date",
                                    "phones"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "phones": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "(99) 9999-9999"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contato criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entidade não processável",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Retorna um contato específico pelo ID",
                "operationId": "2a2c193f37e667891b7a6bc95e8e4132",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contato não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Atualiza um contato existente pelo ID",
                "operationId": "45636eb97c78bd0e474d5507fde18f1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do contato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "birth_date",
                                    "phones"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "phones": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "(99) 9999-9999"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contato atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contato não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Exclui um contato existente pelo ID",
                "operationId": "a4fa6d0e82512d1fac868a7b61e6a2e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contato excluído com sucesso"
                    },
                    "404": {
                        "description": "Contato não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contact": {
                "title": "Contact",
                "description": "Schema para a entidade Contact",
                "properties": {
                    "id": {
                        "description": "ID do contato",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nome do contato",
                        "type": "string"
                    },
                    "email": {
                        "description": "Endereço de e-mail do contato",
                        "type": "string",
                        "format": "email"
                    },
                    "birth_date": {
                        "description": "Data de nascimento do contato (YYYY-MM-DD)",
                        "type": "string",
                        "format": "date"
                    },
                    "phones": {
                        "description": "Lista de números de telefone do contato",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "description": "Data e hora de criação do contato",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data e hora da última atualização do contato",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}